Coloca long long no fast pow

+++ fast_pot.cpp
@@ -4,7 +4,7 @@ using namespace std;
 int mod = 1e9+7;
 int pot (int x, int y) {
 	if (y == 0) return 1;
-	int res = pot(x, y/2);
+	long long res = pot(x, y/2);
 	res = (res * res) % mod;
 	if (y % 2) res = (res * x) % mod;
 	return res;
